app/backend/requirements.txt
asgiref==3.5.2
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==3.4
mypy-extensions==0.4.3
oauthlib==3.2.1
pathspec==0.10.1
platformdirs==2.5.2
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
tomli==2.0.1
tweepy==4.10.1
typing_extensions==4.4.0
urllib3==1.26.12
